Migrated Spring boot 2.6.0 to Spring boot 3.0.0
