+++ src/main/java/com/xss/filter/annotation/XssFilter.java
@@ -1,10 +1,6 @@
-package com.xss.filters.annotation;
+package com.xss.filter.annotation;
 
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * <p>Use this annotation on your controller methods/actions where you wish to filter Cross-site
