+++ src/main/java/com/xss/filter/service/ServletInputStreamXssFilterService.java
@@ -1,0 +1,16 @@
+package com.xss.filter.service;
+
+import jakarta.servlet.ServletInputStream;
+
+/**
+ * ServletInputStreamXssFilterProvider to get ServletInputStream based on supported content type
+ *
+ * @author Bhushan Uniyal
+ */
+public interface ServletInputStreamXssFilterService {
+
+    ServletInputStream getInputStream(ServletInputStream servletInputStream);
+
+    boolean isSupportedContentType(String contentType);
+
+}
